+++ src/components/sidebar/sidebar-menu/SidebarMenu.scss
@@ -32,7 +32,6 @@
   &__navigation {
     display: flex;
     height: 100%;
-    //min-width: 500px;
     align-items: center;
     flex-grow: 1;
     background-color: #151B1F;
@@ -60,6 +59,9 @@
   &__social-network img:hover {
     cursor: pointer;
     background-color: $main-green-color;
+  }
+  &__lang-btn {
+    display: none;
   }
   &__empty {
     width: 688px;
@@ -94,20 +96,28 @@
 @media screen and (min-width: 320px) and (max-width: 767px){
   .sidebar-menu {
     display: block;
-  }
-  .sidebar-menu_open {
-    display: block;
-  }
-  .sidebar-menu__btn-block {
-    display: block;
-    height: 10%;
-    padding-left: 1rem;
-    padding-top: 1rem;
-  }
-  .sidebar-menu__navigation {
-    padding-left: 1rem;
-    height: 90%;
-    align-items: flex-start;
+
+    &_open {
+      display: block;
+    }
+    &__btn-block {
+      display: block;
+      height: 10%;
+      padding-left: 2rem;
+      padding-top: 1rem;
+    }
+    &__navigation {
+      padding-left: 2rem;
+      height: 80%;
+      align-items: flex-start;
+    }
+    &__lang-btn {
+      display: block;
+      height: 10%;
+      background-color: $main-black-color;
+      padding-left: 1rem;
+      color: $main-white-color
+    }
   }
 }
