+++ src/constants/common.ts
@@ -1,4 +1,4 @@
 export const EMPTY_STRING = '';
 
-export const ENG_LANG = 'en';
-export const RU_LANG = 'ru';
+export const ENG_LANG = 'Eng';
+export const RU_LANG = 'Рус';
